Conflicts:
	Dead City II/src/items/RangedWeapon.java

+++ Dead City II/src/sprites/Barrier.java
@@ -8,7 +8,7 @@ public class Barrier extends Sprite {
 
 	private PImage image; 
 	
-	public Barrier(int x, int y, int w, int h) {
+	public Barrier(double x, double y, double w, double h) {
 		super(x, y, w, h);
 		image = DrawingSurface.resources.getImage("Barrier");
 	}
